sdm660: init.qcom.rc: remove interface line from hostapd_fst
The interface line in hostapd_fst is the same as the
default hostapd service and this can cause the lazy
HAL mechanism to start hostapd_fst instead of the
default hostapd, causing various errors.
This change only fixes the errors with default hostapd
operation. In order to fix FST feature with hostapd_fst,
additional changes are needed such as making it use
the default hostapd service and adding sepolicy rules
to support it.

Change-Id: I133336b1ae18fa823969f240b3e4376989fbf017